fix: move add_roll_result_thumb to misc
